sdl mousewheel input event restructure - support both mousewheel axis simultaneously - add EV_GUI_MouseEvent and change to int32 in event_t - clean exisint relating sections

+++ src/common/rendering/v_video.h
@@ -261,7 +261,7 @@ public:
 
 	virtual void PostProcessScene(bool swscene, int fixedcm, float flash, const std::function<void()> &afterBloomDrawEndScene2D) { if (afterBloomDrawEndScene2D) afterBloomDrawEndScene2D(); }
 
-	void ScaleCoordsFromWindow(int16_t &x, int16_t &y);
+	void ScaleCoordsFromWindow(int32_t &x, int32_t &y);
 
 	virtual void Draw2D() {}
